consolidate events
